refactor(shadcn): :package: migrate components to shadcn/ui and consolidate shared components
- Move Button, Label, Card, Accordion, Drawer, DropdownMenu, and Carousel components to shadcn/ui
- Rename Input component to OutsideInput to avoid naming conflicts
- Add hover cursor styles to interactive elements
- Remove duplicate component files from original locations
- Update imports across the application to use new shadcn/ui paths
