+++ app/[locale]/components/atoms/landing/ButtonPrimary.tsx
@@ -1,5 +1,5 @@
 import { useTranslations } from 'next-intl';
-import { Button } from '@atoms/shared/Button';
+import { Button } from '@shadcn/ui/button';
 import Link from 'next/link';
 
 export default function ButtonPrimary() {
@@ -7,7 +7,7 @@ export default function ButtonPrimary() {
 
   return (
     <Link href="/register">
-      <Button className="bg-primary : flex h-[70px] items-center justify-center rounded-full text-2xl font-extrabold text-white max-[580px]:h-[12vw] max-[580px]:w-[35vw] max-[580px]:min-w-[33vw] max-[580px]:text-[4vw]">
+      <Button className="bg-primary flex h-[70px] items-center justify-center rounded-full text-2xl font-extrabold text-white hover:cursor-pointer max-[580px]:h-[12vw] max-[580px]:w-[35vw] max-[580px]:min-w-[33vw] max-[580px]:text-[4vw]">
         {t('buttonStartFree')}
       </Button>
     </Link>
