+++ app/[locale]/components/atoms/shared/OutsideInput.tsx
