+++ app/[locale]/components/shadcn/ui/card.tsx
